SERVER-13201 support $out to foreign database
Allows $out to write to a database foreign to the aggregation namespace
when the mode is "insertDocuments" or "replaceDocuments"

+++ jstests/sharding/out_to_non_existing.js
@@ -2,20 +2,19 @@
 (function() {
     "use strict";
 
-    const st = new ShardingTest({shards: 2, rs: {nodes: 1}});
+    load("jstests/aggregation/extras/utils.js");  // For assertErrorCode.
 
-    const mongosDB = st.s0.getDB(jsTestName());
-    const mongosColl = mongosDB[jsTestName()];
-    const mongosTargetColl = mongosDB[jsTestName() + "_out"];
+    const st = new ShardingTest({shards: 2, rs: {nodes: 1}, config: 1});
+    const sourceDB = st.s0.getDB("source_db");
 
+    /**
+     * Run an aggregation on 'sourceColl' that writes documents to 'targetColl' with $out.
+     */
     function testOut(sourceColl, targetColl, shardedSource) {
-        jsTestLog("Testing $out to non-existent target collection (source collection sharded : " +
-                  shardedSource + ").");
         sourceColl.drop();
-        targetColl.drop();
 
         if (shardedSource) {
-            st.shardColl(sourceColl, {_id: 1}, {_id: 0}, {_id: 1}, mongosDB.getName());
+            st.shardColl(sourceColl, {_id: 1}, {_id: 0}, {_id: 1}, sourceDB.getName());
         }
 
         for (let i = 0; i < 10; i++) {
@@ -25,26 +24,71 @@
         // Test the behavior for each of the $out modes. Since the target collection does not exist,
         // the behavior should be identical.
         ["insertDocuments", "replaceDocuments", "replaceCollection"].forEach(mode => {
+            jsTestLog(
+                `Testing $out from ${sourceColl.getFullName()} ` +
+                `(${shardedSource ? "sharded" : "unsharded"}) to ${targetColl.getFullName()} ` +
+                `with mode "${mode}"`);
+
+            // TODO (SERVER-36832): "replaceCollection" doesn't work in a sharded cluster if the
+            // source and target database are different.
+            if (mode === "replaceCollection" && sourceColl.getDB() !== targetColl.getDB()) {
+                assertErrorCode(sourceColl,
+                                [{
+                                   $out: {
+                                       to: targetColl.getName(),
+                                       db: targetColl.getDB().getName(),
+                                       mode: "replaceCollection"
+                                   }
+                                }],
+                                50939);
+                return;
+            }
+
             targetColl.drop();
-            sourceColl.aggregate([{$out: {to: targetColl.getName(), mode: mode}}]);
+            sourceColl.aggregate(
+                [{$out: {db: targetColl.getDB().getName(), to: targetColl.getName(), mode: mode}}]);
             assert.eq(10, targetColl.find().itcount());
         });
 
-        // Test with legacy syntax, which should behave identical to mode "replaceCollection".
-        targetColl.drop();
-        sourceColl.aggregate([{$out: targetColl.getName()}]);
-        assert.eq(10, targetColl.find().itcount());
+        // If 'targetColl' is in the same database as 'sourceColl', test the $out stage using the
+        // legacy syntax, which must behave identically to mode "replaceCollection".
+        if (targetColl.getDB() == sourceColl.getDB()) {
+            jsTestLog(
+                `Testing $out from ${sourceColl.getFullName()} ` +
+                `(${shardedSource ? "sharded" : "unsharded"}) to ${targetColl.getFullName()} ` +
+                `with legacy syntax`);
+
+            targetColl.drop();
+            sourceColl.aggregate([{$out: targetColl.getName()}]);
+            assert.eq(10, targetColl.find().itcount());
+        }
     }
 
-    //
-    // Test with unsharded source collection to a non-existent target collection.
-    //
-    testOut(mongosColl, mongosTargetColl, false);
+    const sourceColl = sourceDB["source_coll"];
+    const outputCollSameDb = sourceDB["output_coll"];
 
-    //
-    // Test with sharded source collection to a non-existent target collection.
-    //
-    testOut(mongosColl, mongosTargetColl, true);
+    // Test $out from an unsharded source collection to a non-existent output collection in the same
+    // database.
+    testOut(sourceColl, outputCollSameDb, false);
 
+    // Like the last test case, but perform a $out from a sharded source collection to a
+    // non-existent output collection in the same database.
+    testOut(sourceColl, outputCollSameDb, true);
+
+    // Test that $out in a sharded cluster fails when the output is sent to a different database
+    // that doesn't exist.
+    const foreignDb = st.s0.getDB("foreign_db");
+    const outputCollDiffDb = foreignDb["output_coll"];
+    foreignDb.dropDatabase();
+    assert.throws(() => testOut(sourceColl, outputCollDiffDb, false));
+    assert.throws(() => testOut(sourceColl, outputCollDiffDb, true));
+
+    // Test $out from an unsharded source collection to an output collection in a different database
+    // where the database exists but the collection does not.
+    assert.commandWorked(foreignDb["test"].insert({_id: "forcing database creation"}));
+    testOut(sourceColl, outputCollDiffDb, false);
+
+    // Like the last test, but with a sharded source collection.
+    testOut(sourceColl, outputCollDiffDb, true);
     st.stop();
 }());
